backend/requirements.txt
asgiref==3.2.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
ipython==7.11.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.15.2
openpyxl==3.0.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.8
